lesson12

+++ src/routes/usersRoutes.ts
@@ -2,13 +2,15 @@ import express, {NextFunction, Request, Response} from "express";
 import {CreateUserModule, DeleteAndGetUserModule, GetUsersModule, IGetUsers} from "../ts/types";
 import {usersRepository} from "../userRepository";
 import {HTTP_STATUSES} from "../utilities";
+import {body, param, query, validationResult} from 'express-validator';
 
 const userRouter = express.Router();
-
+const idParamRequired = param('id').notEmpty().escape()
+const bodyNameRequired = body('name').matches(/^[^0-9]+$/, "g").trim().notEmpty()
 
 userRouter
     // Получение пользователей
-    .get('/', (req: GetUsersModule<{ name: string }>, res: Response<IGetUsers[]>) => {
+    .get('/',query('name').isString, (req: GetUsersModule<{ name: string }>, res: Response<IGetUsers[]>) => {
         const getUsers = usersRepository.giveUsers(req.query.name)
         if (getUsers.status === 200) {
             // @ts-ignore
@@ -21,7 +23,7 @@ userRouter
     })
 
     // Получение определенного пользователя. :param - параметр
-    .get('/:id', (req: DeleteAndGetUserModule<{ id: string }>, res: Response<IGetUsers>) => {
+    .get('/:id',(req: DeleteAndGetUserModule<{ id: string }>, res: Response<IGetUsers>) => {
         const getUserById = usersRepository.giveUserById(req.params.id)
         if(getUserById.status === 200) {
             // @ts-ignore
@@ -34,25 +36,38 @@ userRouter
     })
 
     // Добавление пользователя
-    .post('/', (req: CreateUserModule<{ name: string }>, res: Response) => {
-        const createUser = usersRepository.createUser(req.body.name)
-        res.status(createUser.status).send(createUser.data)
+    .post('/', bodyNameRequired,(req: CreateUserModule<{ name: string }>, res: Response) => {
+        const result = validationResult(req)
+        if (result.isEmpty()) {
+            const createUser = usersRepository.createUser(req.body.name)
+            res.status(createUser.status).send(createUser.data)
+        }
+        res.status(400)
+            .send({errors: result.array()})
     })
 
     // Удаление пользователя
-    .delete('/:id', (req: Request<{ id: string }>, res: Response) => {
-        if (!req.params.id || +req.params.id < 0) {
-            res.sendStatus(HTTP_STATUSES.BadRequest400)
+    .delete('/:id?', idParamRequired, (req, res) => {
+        const result = validationResult(req);
+        if (result.isEmpty()) {
+            // @ts-ignore
+            const deleteUser = usersRepository.deleteUser(req.params.id)
+            res.sendStatus(deleteUser.status)
         }
-        const deleteUser = usersRepository.deleteUser(req.params.id)
-        res.sendStatus(deleteUser.status)
+        res.status(400)
+            .send({errors: result.array()})
     })
 
     // Обновление данных у пользователя
-    .put('/:id', express.json(), (req: Request<{ id: string }>, res: Response) => {
-        const updateUser = usersRepository.updateUser(req.params.id, req.body.name)
-        res.status(updateUser.status)
-            .send(updateUser.data)
+    .put('/:id?', idParamRequired, bodyNameRequired,(req: Request<{ id: string }>, res: Response) => {
+        const result = validationResult(req);
+        if (result.isEmpty()) {
+            const updateUser = usersRepository.updateUser(req.params.id, req.body.name)
+            res.status(updateUser.status)
+                .send(updateUser.data)
+        }
+        res.status(400)
+            .send({errors: result.array()});
     })
 
 export default userRouter
